+++ examples/data-sources/iosxr_evpn_group/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_evpn_group" "example" {
+  group_id = 1
+}