+++ examples/resources/iosxr_evpn_group/resource.tf
@@ -1,0 +1,8 @@
+resource "iosxr_evpn_group" "example" {
+  group_id = 1
+  core_interfaces = [
+    {
+      interface_name = "Bundle-Ether111"
+    }
+  ]
+}